+++ src/components/NavigationDrawer/types.ts
@@ -1,0 +1,7 @@
+import { DrawerProps } from '@mui/material'
+
+export interface Props {
+  open: boolean
+  variant: DrawerProps['variant']
+  onClose: () => void
+}
